requirements-dev.txt
asgiref==3.5.0
autopep8==1.6.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
defusedxml==0.7.1
Deprecated==1.2.13
dj-database-url==0.5.0
dj-static==0.0.6
Django==4.0.3
django-cors-headers==3.11.0
django-oauth-toolkit==1.7.0
django-rest-passwordreset==1.2.1
djangorestframework==3.13.1
drf-social-oauth2==1.2.1
ecdsa==0.17.0
idna==3.3
itypes==1.2.0
Jinja2==3.0.3
jwcrypto==1.0
MarkupSafe==2.1.0
oauth==1.0.1
oauth2-provider==0.0
oauthlib==3.2.0
pyasn1==0.4.8
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
python-decouple==3.6
python-jose==3.3.0
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
static3==0.7.0
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.8
whitenoise==6.0.0
wrapt==1.13.3
